split out MainGfxWindow class into its own file
